changed int to float

+++ calculator.py
@@ -11,29 +11,29 @@ if input('y or n -- ') == 'y':
     x = input()
 
     if(x) == '1' :
-            a = int(input('#1 -- '))
-            b = int(input('#2 -- '))
+            a = float(input('#1 -- '))
+            b = float(input('#2 -- '))
     
             print('--------- + ')
             print(a + b)
 
     if(x) == '2' :
-            c = int(input('#1 -- '))
-            d = int(input('#2 -- '))
+            c = float(input('#1 -- '))
+            d = float(input('#2 -- '))
     
             print('--------- - ')
             print(c - d)
 
     if(x) == '3' :
-            e = int(input('#1 -- '))
-            f = int(input('#2 -- '))
+            e = float(input('#1 -- '))
+            f = float(input('#2 -- '))
     
             print('---------- *')
             print(e * f)
 
     if(x) == '4' :
-            g = int(input('#1 -- '))
-            h = int(input('#2 -- '))
+            g = float(input('#1 -- '))
+            h = float(input('#2 -- '))
     
             print('----------- /')
             print(g / h)
